move `tools/*` to `chia/dev/*`
