+++ .github/workflows/upload-pypi-source.yml
@@ -128,9 +128,9 @@ jobs:
           - name: black
             command: black --check --diff .
           - name: flake8
-            command: flake8 benchmarks build_scripts chia tests tools *.py
+            command: flake8 benchmarks build_scripts chia tests *.py
           - name: pylint
-            command: pylint benchmarks build_scripts chia tests tools *.py
+            command: pylint benchmarks build_scripts chia tests *.py
           - name: generated protocol tests
             command: |
               python3 -m tests.util.build_network_protocol_files
