+++ chia/dev/analyze-chain.py
