+++ chia/dev/run_benchmark.sh
@@ -4,12 +4,12 @@
 
 run_benchmark() {
    # shellcheck disable=SC2086
-   python -m tools.test_full_sync run $3 --profile --test-constants "$1" &
+   python -m chia.dev.test_full_sync run $3 --profile --test-constants "$1" &
    test_pid=$!
-   python -m tools.cpu_utilization $test_pid
+   python -m chia.dev.cpu_utilization $test_pid
    mkdir -p "$2"
    mv test-full-sync.log cpu.png cpu-usage.log plot-cpu.gnuplot "$2"
-   python -m tools.test_full_sync analyze
+   python -m chia.dev.test_full_sync analyze
    mv slow-batch-*.profile slow-batch-*.png "$2"
    # python -m chia.util.profiler profile-node >"$2/node-profile.txt"
    # mv profile-node "$2"
@@ -24,8 +24,8 @@ else
 fi
 
 # generate the test blockchain databases by running:
-# python -m tools.generate_chain --fill-rate 0 --length 1500 --block-refs 1
-# python -m tools.generate_chain --fill-rate 100 --length 500 --block-refs 0
+# python -m chia.dev.generate_chain --fill-rate 0 --length 1500 --block-refs 1
+# python -m chia.dev.generate_chain --fill-rate 100 --length 500 --block-refs 0
 run_benchmark stress-test-blockchain-1500-0-refs.sqlite "${TEST_NAME}-sync-empty" ""
 run_benchmark stress-test-blockchain-1500-0-refs.sqlite "${TEST_NAME}-keepup-empty" --keep-up
