+++ chia/dev/test_full_sync.py
@@ -19,6 +19,7 @@ import zstd
 
 from chia.cmds.init_funcs import chia_init
 from chia.consensus.default_constants import DEFAULT_CONSTANTS
+from chia.dev.test_constants import test_constants as TEST_CONSTANTS
 from chia.full_node.full_node import FullNode
 from chia.server.outbound_message import Message, NodeType
 from chia.server.server import ChiaServer
@@ -29,7 +30,6 @@ from chia.types.full_block import FullBlock
 from chia.types.peer_info import PeerInfo
 from chia.util.config import load_config
 from chia.util.ints import uint16
-from tools.test_constants import test_constants as TEST_CONSTANTS
 
 
 class ExitOnError(logging.Handler):
