+++ mypy-exclusions.txt
@@ -1,6 +1,9 @@
 # File created by: python manage-mypy.py build-exclusions
 chia.cmds.start_funcs
 chia.daemon.server
+chia.dev.analyze-chain
+chia.dev.run_block
+chia.dev.test_full_sync
 chia.introducer.introducer
 chia.introducer.introducer_api
 chia.plotters.bladebit
@@ -121,6 +124,3 @@ tests.wallet.test_wallet_blockchain
 tests.wallet.test_wallet_interested_store
 tests.wallet.test_wallet_key_val_store
 tests.wallet.test_wallet_user_store
-tools.analyze-chain
-tools.run_block
-tools.test_full_sync
