+++ tests/build-init-files.py
@@ -64,7 +64,7 @@ def command(verbose, root_str):
     stream_handler = logging.StreamHandler()
     logger.addHandler(stream_handler)
 
-    tree_roots = ["benchmarks", "build_scripts", "chia", "tests", "tools"]
+    tree_roots = ["benchmarks", "build_scripts", "chia", "tests"]
     failed = False
     root = pathlib.Path(root_str).resolve()
     directories = [
