+++ tests/tools/test_run_block.py
@@ -6,11 +6,11 @@ from pathlib import Path
 from typing import List
 
 from chia.consensus.default_constants import DEFAULT_CONSTANTS
+from chia.dev.run_block import run_json_block
 from chia.types.blockchain_format.sized_bytes import bytes32
 from chia.types.condition_opcodes import ConditionOpcode
 from chia.types.condition_with_args import ConditionWithArgs
 from chia.util.ints import uint32, uint64, uint128
-from tools.run_block import run_json_block
 
 constants = dataclasses.replace(
     DEFAULT_CONSTANTS,
